WEA-27 Rename function
